// src/app/model/RecetaMedicamentoDetalle.model.ts
export interface RecetaMedicamento {
  correlativoAtencion: string;
  direccionDuenio: string;
  emailDuenio: string;
  nombreDuenio: string;
  apellidoDuenio: string;
  distritoDuenio: string;
  telefonoDuenio: string;
  indicacionMedicamento: string;
  medicamentoRecetado: string;
  flagExisteMedicamento: string;
  farmacia:string;
}